♻️ Refactor: 이메일 및 닉네임 중복 확인 API 위치변경
- `UserController`로 관련 엔드포인트 이동
- 이메일 및 닉네임 중복 확인 로직 구현
- 기존 `AuthController`에서 해당 엔드포인트 제거
